+++ CHANGELOG.md
@@ -19,7 +19,8 @@
 - Implemented `weakref.sll`, `weakref$weakref_type`, `weakref$NO_OBJECT`, `weakref$get` and `weakref$ref` ([#257])
 - Object creation code to build an array from a pointer and a length: `L`
 - Object creation code to build an unsigned 32-bit integer: `u`
-- Sandbox flag for disabling thread API: `no-thread`
+- Sandbox flag for disabling the random API: `no-random`
+- Sandbox flag for disabling the thread API: `no-thread`
 - String format that converts floating-point values to strings: `%f`
 
 #### Built-in Audit Events